07-end-to-end-project/networkrail_producer/pyproject.toml
[tool.poetry]
name = "e2e"
version = "0.1.0"
description = ""
authors = ["zkan <zkan.cs@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2023.3"
"stomp.py" = "^8.1.0"
psycopg2 = "^2.9.6"
SQLAlchemy = "^2.0.9"
dbt-core = "^1.5.0"
dbt-bigquery = "^1.5.1"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
ruff = "^0.0.269"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
